# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AssurAssistance Public API
  version: "1.0.0"
  summary: Travel insurance issuance, certificates, invoices and reporting.
  description: |
    Programmatic access to AssurAssistance for travel agencies, partner
    insurers, and field-sales tooling. Use this API to:

      * List the travel-insurance plans available to sell.
      * Create travellers and travel cases, then confirm a sale.
      * Download invoice and certificate PDFs.
      * Mark invoices as paid and reconcile.
      * Pull ledger and "invoices by region" reports.
      * Download contractual documentation (Terms & Conditions full / brief);
        Admin JWT uploads and replaces those files.
      * Admin JWT may change a supervisor / agency login email
        (`PATCH /admin/agents/{id}` + `GET /admin/email-available`); the address
        must be unique among all login accounts.

    The same JSON shapes are used by the in-house AssurAssistance web app, so
    anything the web app can do, a 3rd-party integration can do — within the
    permission scopes granted to its API key.

  contact:
    name: AssurAssistance Integration Support
    email: support@assurassistancepro.org
  license:
    name: Proprietary
servers:
  - url: https://backend-api.assurassistancepro.org/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
...
security:
  - bearerJWT: []
  - bearerApiKey: []
components:
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /auth/login`. **Used by the in-house web app only.**
        3rd-party integrations should use the `bearerApiKey` scheme.
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        API key issued by an AssurAssistance admin via `POST /admin/api-keys`.
        Keys look like `aas_live_…`. Pass them as:

        ```
        Authorization: Bearer aas_live_xxxxxxxxxxxxxxxxxxxxxxxxxxxx
        ```

        Each request must satisfy the scopes the key was issued with (see the
        `scopes` table in the Integration Guide).
  parameters:
...
    Search:
      name: search
      in: query
      schema: { type: string }
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT to send in the Authorization Bearer header.
        user:
          $ref: '#/components/schemas/User'
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Exchange email + password for a JWT.
      description: |
        For 3rd parties, **do not call this** — use an API key from the admin
        instead. The login endpoint is meant for the in-house web app.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "401":
          description: Wrong credentials or revoked account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /catalogue:
    get:
      tags: [Catalogue]
      summary: List available plans.
      security:
        - bearerJWT: []
        - bearerApiKey: [catalogue:read]
      responses:
        "200":
          description: Plan list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Plan' }
    post:
      tags: [Catalogue]
      summary: Create a plan (admin only).
      description: Internal admin endpoint. No API-key scope grants access.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlanRequest' }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  id: { type: integer }
        "500":
          description: Server error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /cases:
    get:
      tags: [Cases]
      summary: List cases visible to the caller.
      description: |
        Visibility follows the caller's role:
          * **admin** — every case.
          * **sub_admin** — cases under agencies the sub_admin owns.
          * **insurer_supervisor** — all cases whose selected plan has matching `partner_insurer`.
          * **agent** — own cases + sub-agents' cases.
      security:
        - bearerJWT: []
        - bearerApiKey: [cases:read]
      responses:
        "200":
          description: Cases array.
    post:
      tags: [Cases]
      summary: Create a single case (one traveller).
      security:
        - bearerJWT: []
        - bearerApiKey: [cases:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  caseId: { type: integer }
        "400":
          description: Validation error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /sales:
    post:
      tags: [Sales]
      summary: Create a sale from a confirmed case (alternative to /cases/{id}/confirm-sale).
      security:
        - bearerJWT: []
        - bearerApiKey: [sales:write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSaleRequest' }
      responses:
        "201": { description: Created. }
    get:
      tags: [Sales]
      summary: List sales visible to the caller.
      security:
        - bearerJWT: []
        - bearerApiKey: [sales:read]
      responses:
        "200":
          description: Sales array.
...
  /ledger:
    get:
      tags: [Reports]
      summary: Paginated sales ledger with commission + consolidation summary.
      description: |
        Available to admin, sub-admin and agent (partner-manager) JWTs, each
        scoped to their visibility tree. Also via API key with `ledger:read`.
        Each row includes `commission` and `net_to_transfer` using the same
        rules as partner invoices. `meta.summary` covers the full filtered set:
        total policies, total commissions, net amount to transfer.
      security:
        - bearerJWT: []
        - bearerApiKey: [ledger:read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/PaymentStatus'
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Ledger page + filter summary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/LedgerRow' }
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                      summary: { $ref: '#/components/schemas/LedgerSummary' }
...
  /invoice-ledger:
    get:
      tags: [Reports]
      summary: Paginated invoices, grouped by region.
      security:
        - bearerJWT: []
        - bearerApiKey: [invoices:read]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/PaymentStatus'
        - name: region
          in: query
          description: Filter to a single region value (country name or destination string).
          schema: { type: string }
        - name: regionBy
          in: query
          description: Which dimension drives the "region" column.
          schema:
            type: string
            enum: [residence, destination, agent]
            default: residence
      responses:
        "200":
          description: Paginated rows + region summary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/InvoiceLedgerRow' }
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                      regionBy: { type: string }
                  regionSummary:
                    type: array
                    items: { $ref: '#/components/schemas/RegionSummaryRow' }
...
  /admin/api-keys:
    post:
      tags: [API keys]
      summary: Mint a new API key.
      description: |
        **Admin only.** Returns the full `secret` exactly once — store it
        securely. Subsequent reads will only show the prefix.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, owner_user_id, scopes]
              properties:
                name: { type: string, example: "Partner Travel Co. — production" }
                owner_user_id:
                  type: integer
                  description: |
                    The user the key acts on behalf of. Inherits that user's
                    visibility scope (a key owned by an agent only sees that
                    agent's cases).
                scopes:
                  type: array
                  items: { type: string }
                  example: [catalogue:read, cases:read, cases:write, sales:read, sales:write]
                ip_allowlist:
                  type: string
                  nullable: true
                  description: Comma-separated list of IPs / CIDRs the key is allowed from.
                rate_limit_per_min:
                  type: integer
                  nullable: true
                  description: Override the default per-key requests-per-minute limit.
                expires_at:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        "201":
          description: Created (full secret returned once).
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/ApiKeyCreated' }
                  message: { type: string }
    get:
      tags: [API keys]
      summary: List API keys.
      security:
        - bearerJWT: []
      parameters:
        - name: owner_user_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Keys array.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ApiKey' }
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Send a 6-digit password-reset OTP by email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: OTP sent. }
        "404":
          description: User not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/verify-reset-code:
    post:
      tags: [Authentication]
      summary: Verify a password-reset OTP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, description: The 6-digit OTP. }
      responses:
        "200": { description: Code verified. }
        "400":
          description: Invalid or expired code.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Set a new password using a verified OTP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, newPassword, confirmPassword]
              properties:
                email: { type: string, format: email }
                code: { type: string, description: The 6-digit OTP. }
                newPassword: { type: string }
                confirmPassword: { type: string }
      responses:
        "200": { description: Password reset. }
        "400":
          description: Invalid OTP or weak/compromised password.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # --- Agents / agency management (JWT: admin or sub-administrator) ----------
...
  /reconciliation:
    get:
      tags: [Reconciliation]
      summary: Monthly agent reconciliation.
      description: |
        Returns one row per issuing account for the selected month/year.
        Includes **travel agency** (`agency_name`) and supervising
        **sub-admin / sales representative** (`sub_admin_name`).
      security:
        - bearerJWT: []
      parameters:
        - name: month
          in: query
          required: true
          description: Three-letter month label (e.g. Jan, Sep).
          schema:
            type: string
            enum: [Jan, Feb, Mar, Apr, May, Jun, Jul, Aug, Sep, Oct, Nov, Dec]
        - name: year
          in: query
          required: true
          schema: { type: integer, example: 2026 }
      responses:
        "200":
          description: Reconciliation rows.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ReconciliationRow' }
        "400":
          description: Missing or invalid month/year.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /activity-log:
    get:
      tags: [Activity log]
      summary: Paginated user-activity audit log (admin only).
      security:
        - bearerJWT: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        "200":
          description: Activity page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { type: object }
                  meta:
                    $ref: '#/components/schemas/Pagination'
...
  /contractual-documents:
    get:
      tags: [Contractual documents]
      summary: List full and brief Terms & Conditions slots.
      security:
        - bearerJWT: []
      responses:
        "200":
          description: Both document slots (has_file false when not uploaded).
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  keys:
                    type: array
                    items: { type: string, enum: [full, brief] }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ContractualDocument' }
...
  /quotes:
    post:
      tags: [Quotes]
      summary: Create and store a quote (case awaiting payment).
      description: |
        Creates the traveller and a `cases` row with status `AwaitingPayment` and a
        customer-facing reference `QT-XXXXXXXX`.

        The premium is RECALCULATED from the live catalogue at this moment: a price
        quoted in a conversation hours earlier is never trusted.

        The resulting case is an ordinary case — `POST /cases/{caseId}/confirm-sale`
        issues the policy, certificate and invoice from it unchanged.

        Inputs are normalised: passport uppercased and stripped of spaces/dashes,
        email lowercased, phone converted to E.164.
      security:
        - bearerJWT: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuoteCreateRequest' }
      responses:
        "201":
          description: Quote stored.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Quote created. Confirm the sale to issue a policy." }
                  data:
                    type: object
                    properties:
                      quote_reference: { type: string, example: "QT-BF664C18" }
                      case_id: { type: integer, example: 512 }
                      traveller_id: { type: integer, example: 388 }
                      status: { type: string, example: AwaitingPayment }
                      pricing: { $ref: '#/components/schemas/QuotePricing' }
        "400":
          description: Validation failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationError' }
        "409":
          description: |
            attribution_missing — no owning account is configured for this source.
            Every case belongs to a user because commissions, partner invoices, the
            ledger and reconciliation are keyed to one.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        "422":
          description: age_ineligible or no_price.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [Quotes]
      summary: List stored quotes.
      description: An API key sees only quotes owned by its own account.
      security:
        - bearerJWT: []
        - bearerApiKey: []
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: source, in: query, schema: { type: string, enum: [web, whatsapp, api] } }
        - { name: status, in: query, schema: { type: string, enum: [Draft, AwaitingPayment, Confirmed, Cancelled] } }
      responses:
        "200":
          description: Paginated quotes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { type: object } }
                  pagination: { type: object }
...
  /zones:
    get:
      tags: [Zones]
      summary: Zones in use, with country counts.
      security:
        - bearerJWT: []
        - bearerApiKey: []
      responses:
        "200":
          description: Zones.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        zone: { type: string, example: Worldwide }
                        countryCount: { type: integer, example: 195 }
                        activeCount: { type: integer, example: 195 }
...
  /whatsapp/webhook:
    get:
      tags: [WhatsApp]
      summary: Meta verification handshake.
      description: Echoes `hub.challenge` as plain text when `hub.verify_token` matches the configured token. No bearer auth.
      security: []
      parameters:
        - { name: hub.mode, in: query, schema: { type: string, example: subscribe } }
        - { name: hub.verify_token, in: query, schema: { type: string } }
        - { name: hub.challenge, in: query, schema: { type: string } }
      responses:
        "200":
          description: The challenge, echoed.
          content:
            text/plain:
              schema: { type: string }
        "403":
          description: Token mismatch, wrong mode, or no token configured.
    post:
      tags: [WhatsApp]
      summary: Inbound messages and delivery receipts from Meta.
      description: |
        Authenticated by `X-Hub-Signature-256` (HMAC-SHA256 of the RAW body with the
        Meta app secret), not by a bearer token.

        Returns `403` when the signature is absent or does not verify. Returns `200`
        in every other case, including internal failures: a non-2xx answer makes Meta
        retry with growing delay and eventually disable the webhook, so a database
        problem must not cost the integration.

        Redelivery is safe — messages are deduplicated on Meta's `wa_message_id`.
        Concurrent messages from one number are serialised, so a customer sending
        three messages in two seconds cannot advance the flow three times.
      security: []
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema: { type: string, example: "sha256=9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Meta webhook envelope. }
      responses:
        "200":
          description: Accepted (also returned on internal failure, by design).
        "403":
          description: Signature missing or invalid.
...